LIVE-4133: extracted HLS types into their own classes. Added domain modeling relationships and classes. Updated documentation.
