+++ docs/hls-parser-expanded.drawio
@@ -1,0 +1,1 @@
+<mxfile host="app.diagrams.net" modified="2022-04-20T14:50:06.943Z" agent="5.0 (Macintosh)" etag="4X248i5ZmmwDagcXc8Sy" version="17.4.5" type="device"><diagram id="iJJovRhbEwnwrPdhqQHO" name="Page-1">7V1bc5s4FP41fkwGcbP9mDhNO7PJbibJdNunHcUoRhOMPCDn0l+/EpIwSMImjrGdlplOaw5CgM6nc/mORAfeZP76NYOL+JpEKBm4TvQ68C4Grgvc8Yj9wyVvQjLyAiGYZTiSjVaCO/wLSaEjpUscobzWkBKSULyoC6ckTdGU1mQwy8hLvdkjSep3XcAZMgR3U5iY0n9xRGP5FoGzkn9DeBarOwNHnplD1VgK8hhG5KUi8r4MvElGCBW/5q8TlPDBU+MirrtsOFs+WIZS2uYC5+///oqd7xicXD9d/iQUPH77fiJ7eYbJUr7wLYqWaQRTekczBOcDNxy43uvZwDuXr0Hf1NjkL3iewJQdnT8SfoE4A9gxTPAsZb+n7OFQxgTPKKOYDeuZPEHJgkmnMU6iK/hGlvwVcgqnT+roPCYZ/sW6hYnsk53OqEQIe6pqizt+JRM7TJqhnLW5UeMCNNE1fK01vII5lYIpSRK4yPFD+RpzmM1wek4oJXPZ6BEnyYQkJCvGwIsCNIr84uky8oQqZ0bugxeG5bhVFaVGnY0Jeq2IpOK+IjJHNHtjTeTZUGJITqKhPHxZIRKMpSyuoHEkJx+Uk2BWdlze65ZNGpjO2BuXN/P8oHY313I3x3I3jpTq3WDCVJ9Cis4JQ1ReBSf7UXnNlaiA7Dvg6xrwHbhsvB0xq87Yr2umXpTdJPAtwUzROoTZ+FNDeSkRmK5oWooUrBP0SBtBnS/gFKezq6LNhb+S3Mqx4iLCrn1MCnsQ4yhCKQckoZBCgT4OtQXBKS0GMzhnf9iQT5zTYBCwB5+wY7A6Zn9484xOSMreBeICaYi9+wvi8LZgcK052AzMt7rONyHR9ZuRWAPFexEARg0QyIXf4Bg4yzLI7hQmfCR0QIQzarNtPTA+BoygpYnynI6A4Vlwoek4wYXuhI6VdwdbKXjOVJWglUbvucIvToChdc/UumfRcAIfUHJDckwx4f1noq2m+UMpd9TW/3Sk27Gh2w1m/o+MVHYQd/jO5sAjtCk+7EjzKi0wzP0zzDAspoZh8b+LU72pb2cNxq1BsiYGsIFCjw93BgrXZus5KKbLjL+sAkABjRINPQp2ioLAPTAKgJnL9h5/V9odjloa/q6muBnmFzmermCURmec9WFHZFFMpZjOlVtmJ6VTBq7y0qp1hOGcpNF9jFPNgQNfCS4xf+hCWyiaIRVFMHXEZEZSmHxZSZl+WdKLItX+FdMf6jHY759czmasOLp4rTS7eFMHKRu5H9WDylX8cHVZcaSu2yFDkZNlNkXr9CLasdGZobUhpsQKH7e1eMtQAil+rrNva/iLGz49VgFLGZqWvknLL8QbyauqbJnWke9v6Ei8stHRrpgNzxboiHgmws8qlmHjE8I5dzXpQ74oNKkfO+qyh2wVAQkJe7BKX+ZcYrq64lZJRaAVa1WZVcpOKoNld4UtDZOa55LOlXcclBxU+9BVQqc1nj6WbJq+R/APvX06rH1S9NBGA6ViyK4NlOfV7Yo/3NZAjTd01LWBMoNucHqatjJIR2Bo1Iz9bIbGjIOaEpqe89iO8/CGWvnDM0mPsjxZjX1LCmz3WjcJL5HgLjMsslpmO1l+YWCgT2rts+cj3IZV951ltUNbVstVv1BcpyhzIRYgvK/KVXWwTS65B0QLmmO/gPD9BkDAZYSJhf68ZWGlIBV6ArQjUFjYkT2DImwAxTPLQnpQHAQUwPEPjYphAyry5QPFNLHWx3tkdI+MthFld8hoWjUxTUiOoglccAT08DgMPIJDB52+jXzsSym70u+oZQzZVTHFdy3qNbjl+xjnvCf+FxRMyTNKMUqnfK0gfVugJh7Z6IrGkKqulmnBlKCId0LYX9+u7sQ450UG68CUn8pRhhlsfkFpa7ajsPMYLlCBQ8rx9RJjiu4YgLjsJYOLOsH0AKdPs4IX/WdJBb4LeQSzp3847GhBYZ46nOHMJfs62i2jufWaUGD6FOuanKArksJvseKqQn9L0xDBPC5YaFDXRZ2ebhyujQSwwvrmCtWgHf9bGd9gzfB+lCbW2N3A2ZYm1upYvq7/jmli30ZdfZKiyPaoG7cEnXIeXVcdANBQALaEk251jI46hlNg0mGi6rCmpGApITQS32UootchzNihZXygJsDHCw3O6dDxhjsxLxocyrUzqgfy+JijbhRoxh7VlKJw6vfN/ryvaWwVK+jrOP2xyU1Y1/MA4HYULgRNi/aKyLJIPO8N/fe5pX0yfZ61mkETU9mXsjpW/cEXaA6bKhd9KesggGi7prM7W2ADRE8z7Ui9tvrDXtfsBrbyw2+f+nltV5wpjXad+w212M/TGYC2uZ/eke9oyOk69zOLFs5x535qBhx37hfUe+gw97Ps21vr7Pt0b7t0L9RWsIUWbnhk8wRdeXr1AH3U/8EoYCeb82yq7yzIC5vy/BzN5si+Y/NOnOpLzF2hwpYL7hcV/Wa9DtVry+ys9r4r7QaGdssp3bv4nbh4bZF6WebblOzp29h2p/Oe2NvJbC/nzufhdMMWiX5vzLdVb1vetjNj/kdW8JXT3UjjKIV2TeMY+/38LWkcvaNA8YR7onGGZi545DSOmgHHRePUA4AToGmxOxpn6BoK/DQGQe1zVr/b7HN+/+q+cmc0GLxrZ/T25kptE9tsrvb0IQagcVCeWvT6XnOld+Truyy7NleWfc5HbKvU5DwqW+Vqpsrbm6n6xCWoT2F2/LZmZ1+fVxhpZkevbrRe6BhoHfnas3Rtdsxi11GbnWOsdJ14dcOzR7tj5kz3Mcr5xgk2kgNtM0XT5obKzgZIaYYflpR/UuYS5jmZYii2aL1zD4a4PYMByYotGNFSkDL6hgu5DUM0T/mXtpwFWSzVzo3ysnLfRr9Dw1KG0wvmlhUZw33W4Ua2OlyzP3zXHo0DeJ+2ObqKjo9kcwcIhqdOCEZBEA4d4KhP0UiUbPvFMhCMT4HrDQHwfMd1QvVFI+2bvXvyX6MWBZ7fEWn+USGNgeF07I6Bx+KikRsqL11u3tA4g7ZQC8YboqytwcUOV/8jhmi++n9FvC//Aw==</diagram></mxfile>
