+++ src/types/ByteRange.ts
@@ -1,0 +1,6 @@
+export interface ByteRange {
+    length?: number;
+    offset?: number;
+}
+
+export default ByteRange;
