+++ src/types/Resolution.ts
@@ -1,0 +1,6 @@
+export interface Resolution {
+    width: number;
+    height: number;
+}
+
+export default Resolution;
